refactor(cli/audio): split resampler submodule into file
